Fix various bugs in stdio_print, stdmem_copy, sys_write, sys_read. exit() now appears to work properly

+++ os/src/kernel/file.h
@@ -3,10 +3,9 @@
 
 #include <stddef.h>
 #include <stdtypes.h>
-#include <sys/stat.h>
-#include <sys/types.h>
+#include <stdio.h>
 
-ssize_t sys_write(int fd, char *ptr, size_t len);
-ssize_t sys_read(int fd, char *ptr, size_t len);
+int32_t sys_write(int fd, char *ptr, size_t len);
+int32_t sys_read(int fd, char *ptr, size_t len);
 
 #endif
